+++ examples/client/src/gen4/getInventory.ts
@@ -1,0 +1,4 @@
+export const getInventory = {
+  method: 'GET',
+  url: '/store/inventory',
+}