+++ packages/plugin-client/src/generators/__snapshots__/importPath/findPetsByTags.ts
@@ -1,0 +1,27 @@
+/* eslint-disable no-alert, no-console */
+import fetch from 'axios'
+import type { FindPetsByTagsQueryTags, FindPetsByTagsQueryStatus, FindPetsByTagsResponse } from './FindPetsByTags'
+import type { Client, RequestConfig, ResponseErrorConfig } from 'axios'
+
+export function getFindPetsByTagsUrl() {
+  const res = { method: 'GET', url: `/pet/findByTags` as const }
+  return res
+}
+
+/**
+ * {@link /pet/findByTags}
+ */
+export async function findPetsByTags(
+  params: { tags: FindPetsByTagsQueryTags; status?: FindPetsByTagsQueryStatus },
+  config: Partial<RequestConfig> & { client?: Client } = {},
+) {
+  const { client: request = fetch, ...requestConfig } = config
+
+  const res = await request<FindPetsByTagsResponse, ResponseErrorConfig<Error>, unknown>({
+    method: 'GET',
+    url: getFindPetsByTagsUrl().url.toString(),
+    params,
+    ...requestConfig,
+  })
+  return res.data
+}