+++ packages/ast/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @kubb/ast
 
+## 5.0.0-alpha.29
+
+## 5.0.0-alpha.28
+
 ## 5.0.0-alpha.27
 
 ### Minor Changes