+++ packages/plugin-msw/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @kubb/plugin-msw
 
+## 5.0.0-alpha.29
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kubb/core@5.0.0-alpha.29
+  - @kubb/oas@5.0.0-alpha.29
+  - @kubb/plugin-faker@5.0.0-alpha.29
+  - @kubb/plugin-oas@5.0.0-alpha.29
+  - @kubb/plugin-ts@5.0.0-alpha.29
+
+## 5.0.0-alpha.28
+
+### Patch Changes
+
+- Updated dependencies [[`d46e725`](https://github.com/kubb-labs/kubb/commit/d46e7255c2419e412ace2e090205d552a885c6ca)]:
+  - @kubb/plugin-ts@5.0.0-alpha.28
+  - @kubb/plugin-faker@5.0.0-alpha.28
+  - @kubb/core@5.0.0-alpha.28
+  - @kubb/oas@5.0.0-alpha.28
+  - @kubb/plugin-oas@5.0.0-alpha.28
+
 ## 5.0.0-alpha.27
 
 ### Patch Changes