+++ packages/plugin-redoc/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @kubb/plugin-redoc
 
+## 5.0.0-alpha.29
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kubb/adapter-oas@5.0.0-alpha.29
+  - @kubb/core@5.0.0-alpha.29
+
+## 5.0.0-alpha.28
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kubb/adapter-oas@5.0.0-alpha.28
+  - @kubb/core@5.0.0-alpha.28
+
 ## 5.0.0-alpha.27
 
 ### Patch Changes